pyproject.toml
[project]
name = "inboxzero"
version = "0.0.0"
description = "A minimal IMAP/SMTP terminal client designed for fast, keyboard-driven email triage.Features include new-email notifications, inbox listing, and a lightweight message composer."
readme = "README.md"
authors = [
    { name="madushanka rajapakshe", email="MadushankaRajapaksha999@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "rich>=10.0.0",
    "typer>=0.9.0",
    "textual>=0.1.0",
    "python-dotenv>=1.0.0",
]

# [project.optional-dependencies]
# # Enhanced notifications (optional fallback if native methods fail)
# notifications = [
#     "plyer>=2.0.0"
# ]
# # All optional features
# all = [
#     "plyer>=2.0.0"
# ]

[project.scripts]
inbox-zero = "inboxzero.cli:main"
inboxzero = "inboxzero.cli:main"

[project.urls]
Homepage = "https://github.com/MadushankaRajapaksha/Inbox-Zero-"
"Bug Tracker" = "https://github.com/MadushankaRajapaksha/Inbox-Zero-/issues"

[tool.setuptools]
packages = ["inboxzero"]

